Extract VideoPlayer component from Home
